Change developer to customisation
Because there are other forms of customisation that can be covered, I thought
that renaming the section would be appropriate. I've split the PDF information
in to its own section as part of that.
I haven't based this on my existing documentation branch so there is likely to
be some conflicts when they are merged.