[Issue #358] Fix: README and documentation for version 1.14
Changes include updating the current README file to reflect the update from version 1.12 to version 1.14 of koalixcrm. A new detailed update guide has been added to the documentation, providing users with specific instructions for this version update, along with advice for backup and installation of Docker to support the new version.